Move rocket list component to components module
