+++ apps/fly-my-rockets/src/app/rockets/components/components.module.ts
@@ -4,10 +4,11 @@ import { SharedModule } from '../../shared/shared.module';
 import { PipesModule } from '../pipes/pipes.module';
 
 import { FlightComponent } from './flight/flight.component';
+import { RocketListComponent } from './rocket-list/rocket-list.component';
 
 @NgModule({
-  declarations: [FlightComponent],
+  declarations: [FlightComponent, RocketListComponent],
   imports: [CommonModule, SharedModule, PipesModule],
-  exports: [FlightComponent]
+  exports: [FlightComponent, RocketListComponent]
 })
 export class ComponentsModule {}
