+++ apps/fly-my-rockets/src/app/rockets/rockets-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { RocketListComponent } from './rocket-list/rocket-list.component';
+import { RocketListComponent } from './components/rocket-list/rocket-list.component';
 import { RocketShowComponent } from './rocket-show/rocket-show.component';
 
 const routes: Routes = [
