+++ apps/fly-my-rockets/src/app/rockets/rockets.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 
 import { RocketsRoutingModule } from './rockets-routing.module';
 import { SharedModule } from '../shared/shared.module';
-import { RocketListComponent } from './rocket-list/rocket-list.component';
 import { RocketDialogComponent } from './dialogs/rocket-dialog/rocket-dialog.component';
 import { RocketShowComponent } from './rocket-show/rocket-show.component';
 import { FlightDialogComponent } from './dialogs/flight-dialog/flight-dialog.component';
@@ -12,7 +11,6 @@ import { ComponentsModule } from './components/components.module';
 
 @NgModule({
   declarations: [
-    RocketListComponent,
     RocketDialogComponent,
     RocketShowComponent,
     FlightDialogComponent
